Add v_interlock to the vnode structure
FreeBSD implements this as 'struct mtx v_interlock' under the vnode structure.
Since 'struct mtx' doesn't seem to be a thing in XNU we are basing this off
of its implementation in Darwin 0.3, which is simple_lock_data_t v_interlock,
because the simple_lock_data_t type IS present in XNU under the private API.
The necessary headers were pulled from xnu-7195.121.3 (latest as of this moment).

No idea if this is going to work as intended but here goes.